単位の変更

+++ app/src/main/java/jp/ac/jec/cm0120/pittan/ui/add_data/AddDataActivity.java
@@ -175,8 +175,8 @@ public class AddDataActivity extends AppCompatActivity {
       photoBitmap = PictureIO.outputPicture(tempPath);
       imageViewPhoto.setVisibility(View.VISIBLE);
       imageViewPhoto.setImageBitmap(photoBitmap);
-      editHeightSize.setText(String.valueOf(mIntent.getDoubleExtra(AppConstant.Objection.EXTRA_MODEL_SIZE_HEIGHT,0.0f)));
-      editWidthSize.setText(String.valueOf(mIntent.getDoubleExtra(AppConstant.Objection.EXTRA_MODEL_SIZE_WIDTH,0.0f)));
+      editHeightSize.setText(String.valueOf(mIntent.getIntExtra(AppConstant.Objection.EXTRA_MODEL_SIZE_HEIGHT,0)));
+      editWidthSize.setText(String.valueOf(mIntent.getIntExtra(AppConstant.Objection.EXTRA_MODEL_SIZE_WIDTH,0)));
     }
   }
